Change naming scheme for kubectl binaries
Do not use the `kubectl-<Maj>.<Min>.<Patch>` format, this can lead to
confusion with kubectl plugins (which are named `kubectl-<plugin>`).

This wouldn't be a problem for local kubectl binaries (as in the ones
under user's home directory), but it would be an issue for system-wide
ones (under `/usr/bin`).

Instead use the `kubectl<Maj>.<Min>.<Patch>` format.