Fix infinite recursion loop with kube login
Problem
---

As raised in #37 and in rancher-sandbox/rancher-desktop#1308,
When using kubelogin or any other credentials benefiting from kubectl
plugins, an infinite recursion loop happens.

kuberlr is designed to be a built-in replacement for `kubectl` and hence
replaces the `kubectl` command in the `PATH`.

Using kubectl exec plugin configured as below,
the Kubernetes version client will systematically
execute kubectl to obtain credentials to the cluster.

This will result calling `kuberlr` which will try to get the version
again.

```yaml
  - name: schip-oidc
    user:
      exec:
        apiVersion: client.authentication.k8s.io/v1beta1
        args:
          - oidc-login
          - get-token
          - --oidc-issuer-url=https://auth.example.com
          - --oidc-client-id=kubelogin
        command: kubectl
```

To prevent this infinite loop, mark the execution of `KubectlVersionToUse`
in an environment variable.

This way, when `kuberlr` is executed again to authenticate to Kubernetes
so it can retrieve the Kubernetes version, it will systematically use
the latest stable Kubectl version and avoid the infinite recursion loop.

fixes: #37
fixes: rancher-sandbox/rancher-desktop#1308